refactor: views 수정
